Activate dictionaries based on document language
This commit adds basic support for selective activation of the
dictionaries based on the language declared by the currently opent
document.

This feature is aimed at users who read documents in different languages
and have a certain number of dictionaries. When enabling the
corresponding option in the dictionary menu, the ReaderDictionary module
automatically deactivates the dictionaries which don't match the
document language.